eliminate inconsistencies identified during a code review.
