+++ stl/inc/ostream
@@ -454,14 +454,14 @@ public:
         return *this;
     }
 
-#if _HAS_CXX20 // P1147R1 "Printing volatile Pointers"
+#if _HAS_CXX20
     template <class = void> // TRANSITION, ABI
     basic_ostream& operator<<(const volatile void* _Val) {
         return *this << (const_cast<const void*>(_Val));
     }
 #endif // _HAS_CXX20
 
-#if _HAS_CXX17 // LWG-2221 "No formatted output operator for nullptr"
+#if _HAS_CXX17
     template <class = void> // TRANSITION, ABI
     basic_ostream& operator<<(nullptr_t) { // insert a null pointer
         return *this << "nullptr";
