+++ tests/std/tests/P1147R1_printing_volatile_pointers/env.lst
