+++ tests/std/tests/P1147R1_printing_volatile_pointers/test.cpp
@@ -1,0 +1,29 @@
+// Copyright (c) Microsoft Corporation.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
+#include <cassert>
+#include <cstddef>
+#include <sstream>
+
+using namespace std;
+
+template <typename T>
+std::string getTextValue(T* ptr) {
+    ostringstream out;
+    out << ptr;
+    return out.str();
+}
+
+void test(int value) {
+    int* p0          = &value;
+    volatile int* p1 = p0;
+
+    const string expected = getTextValue(p0);
+    const string actual   = getTextValue(p1);
+
+    assert(expected == actual);
+}
+
+int main() {
+    test(42);
+}
